:art: Added dart doc to new classes. Restructured some files and updated some values which could be private
