+++ chopper/lib/chopper.dart
@@ -11,9 +11,8 @@ export 'src/chopper_log_record.dart';
 export 'src/constants.dart';
 export 'src/extensions.dart';
 export 'src/chain/chain.dart';
-export 'src/interceptor.dart' hide InternalInterceptor;
-export 'src/http_logging_interceptor.dart';
-export 'src/interceptor.dart';
+export 'src/interceptors/interceptor.dart';
+export 'src/converters.dart';
 export 'src/request.dart';
 export 'src/response.dart';
 export 'src/utils.dart' hide mapToQuery;
