+++ chopper/lib/src/base.dart
@@ -4,7 +4,8 @@ import 'package:chopper/src/annotations.dart';
 import 'package:chopper/src/authenticator.dart';
 import 'package:chopper/src/chain/call.dart';
 import 'package:chopper/src/constants.dart';
-import 'package:chopper/src/interceptor.dart';
+import 'package:chopper/src/converters.dart';
+import 'package:chopper/src/interceptors/interceptor.dart';
 import 'package:chopper/src/request.dart';
 import 'package:chopper/src/response.dart';
 import 'package:http/http.dart' as http;
