+++ chopper/lib/src/interceptors/http_call_interceptor.dart
@@ -1,19 +1,24 @@
 import 'package:chopper/src/chain/chain.dart';
-import 'package:chopper/src/interceptor.dart';
+import 'package:chopper/src/interceptors/internal_interceptor.dart';
 import 'package:chopper/src/response.dart';
 import 'package:http/http.dart' as http;
 
 import '../utils.dart';
 
+/// {@template HttpCallInterceptor}
+/// Internal interceptor that handles the actual HTTP calls. HTTP calls are handled by [_httpClient] for http package.
+/// {@endtemplate}
 class HttpCallInterceptor implements InternalInterceptor {
-  const HttpCallInterceptor(this.httpClient);
+  /// {@macro HttpCallInterceptor}
+  const HttpCallInterceptor(this._httpClient);
 
-  final http.Client httpClient;
+  /// HTTP client to be used for making the actual HTTP calls.
+  final http.Client _httpClient;
 
   @override
   Future<Response<BodyType>> intercept<BodyType>(Chain<BodyType> chain) async {
     final finalRequest = await chain.request.toBaseRequest();
-    final streamRes = await httpClient.send(finalRequest);
+    final streamRes = await _httpClient.send(finalRequest);
 
     if (isTypeOf<BodyType, Stream<List<int>>>()) {
       return Response(streamRes, (streamRes.stream) as BodyType);
