+++ chopper/lib/src/interceptors/internal_interceptor.dart
@@ -1,0 +1,4 @@
+import 'package:chopper/src/interceptors/interceptor.dart';
+
+/// An interface for implementing Internal interceptors only used by Chopper itself.
+abstract interface class InternalInterceptor implements Interceptor {}
