+++ chopper/test/authenticator_test.dart
@@ -1,6 +1,8 @@
 import 'dart:convert' show jsonEncode;
 
 import 'package:chopper/chopper.dart';
+import 'package:chopper/src/converters.dart';
+import 'package:chopper/src/interceptors/header_interceptor.dart';
 import 'package:http/http.dart' as http;
 import 'package:http/testing.dart';
 import 'package:test/test.dart';
