+++ chopper/test/chain/request_converter_interceptor_test.dart
@@ -2,8 +2,9 @@ import 'dart:async';
 
 import 'package:chopper/src/chain/chain.dart';
 import 'package:chopper/src/chain/interceptor_chain.dart';
-import 'package:chopper/src/chain/request_converter_interceptor.dart';
-import 'package:chopper/src/interceptor.dart';
+import 'package:chopper/src/converters.dart';
+import 'package:chopper/src/interceptors/interceptor.dart';
+import 'package:chopper/src/interceptors/request_converter_interceptor.dart';
 import 'package:chopper/src/request.dart';
 import 'package:chopper/src/response.dart';
 import 'package:http/http.dart' as http;
