+++ chopper/test/converter_test.dart
@@ -1,6 +1,7 @@
 import 'dart:convert' as dart_convert;
 
 import 'package:chopper/chopper.dart';
+import 'package:chopper/src/converters.dart';
 import 'package:http/http.dart' as http;
 import 'package:http/testing.dart';
 import 'package:test/test.dart';
