+++ chopper/test/interceptors_test.dart
@@ -1,6 +1,8 @@
 import 'dart:async';
 
 import 'package:chopper/chopper.dart';
+import 'package:chopper/src/interceptors/curl_interceptor.dart';
+import 'package:chopper/src/interceptors/header_interceptor.dart';
 import 'package:http/http.dart' as http;
 import 'package:http/testing.dart';
 import 'package:test/test.dart';
