+++ chopper/test/test_service_base_url.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 import 'dart:convert';
 
 import 'package:chopper/chopper.dart';
+import 'package:chopper/src/converters.dart';
 
 part 'test_service_base_url.chopper.dart';
