+++ chopper/test/test_without_response_service.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 import 'dart:convert';
 
 import 'package:chopper/chopper.dart';
+import 'package:chopper/src/converters.dart';
 import 'package:http/http.dart' show MultipartFile;
 
 part 'test_without_response_service.chopper.dart';
